requirements.txt
fastapi[standard]==0.110.1
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pydantic==2.5.2
python-dotenv==1.0.1
jinja2==3.1.3
numpy==1.26.3
openai>=1.54.0
httpx>=0.27.0,<0.29.0
faiss-cpu==1.7.4
sentence-transformers==2.3.1
alembic==1.13.1
pyyaml==6.0.1
markdown==3.5.2
python-multipart==0.0.9
